Now has a config file on the spiffs that allows everything to be set at "run time", although this file is created and uploaded at build time for now. The intention is to replace the spiffs one with a new one from the SD card once I get that working.
This will report status via a web page, will connect to wifi infrastructure
or if selected or if infrastructure mode fails, it will start a soft access
point and can be connected to this way

The serial test harness is mostly complete.

The sensors are just skeletons at this stage, as is the flight event code.